+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Option to create file directly from the error popup if it doesn't exist
 - `${debugPid}` for PID of the process for debugging
 - Example debug command and attach configurations for C++, Python, and Java as well as convenient ways to automatically add them to launch.json file
+- Toggle show details button for stress tester as well as more statuses for judge
 
 ### Changed